requirements.txt
pandas==2.2.0
pillow==10.2.0
pyodbc==5.1.0